requirements.txt
gradio>=3.50.2
PyYAML>=6.0.1
pywin32>=306